+++ base/file.jl
@@ -8,6 +8,7 @@ export
     chown,
     cp,
     cptree,
+    DirEntry,
     diskstat,
     hardlink,
     mkdir,
@@ -1055,9 +1056,16 @@ const UV_DIRENT_BLOCK = Cint(7)
     DirEntry
 
 A type representing a filesystem entry that contains the name of the entry, the directory, and
-the raw type of the entry. The full path of the entry can be obtained lazily by accessing the
-`path` field. The type of the entry can be checked for by calling [`isfile`](@ref), [`isdir`](@ref),
+the raw type of the entry. The full path of the entry can be obtained lazily via [`path`](@ref Base.Filesystem.path)
+or [`joinpath`](@ref).
+
+Public fields:
+- `dir::String`: The directory containing the entry.
+- `name::String`: The name of the entry.
+
+The type of the entry can be checked for by calling [`isfile`](@ref), [`isdir`](@ref),
 [`islink`](@ref), [`isfifo`](@ref), [`issocket`](@ref), [`ischardev`](@ref), and [`isblockdev`](@ref)
+on the entry object.
 """
 struct DirEntry
     dir::String
@@ -1080,13 +1088,14 @@ isblockdev(obj::DirEntry) = (isunknown(obj) || islink(obj)) ? isblockdev(path(ob
 realpath(obj::DirEntry) = realpath(path(obj))
 
 """
-    _readdirx(dir::AbstractString=pwd(); sort::Bool = true)::Vector{DirEntry}
+    readdir(::Type{DirEntry}, dir::Union{AbstractString,DirEntry}=pwd(); sort::Bool = true) -> Vector{DirEntry}
+    readdir(entry::DirEntry; sort::Bool=true) -> Vector{DirEntry}
 
 Return a vector of [`DirEntry`](@ref) objects representing the contents of the directory `dir`,
 or the current working directory if not given. If `sort` is true, the returned vector is
 sorted by name.
 
-Unlike [`readdir`](@ref), `_readdirx` returns [`DirEntry`](@ref) objects, which contain the name of the
+The [`DirEntry`](@ref) objects that are returned contain the name of the
 file, the directory it is in, and the type of the file which is determined during the
 directory scan. This means that calls to [`isfile`](@ref), [`isdir`](@ref), [`islink`](@ref), [`isfifo`](@ref),
 [`issocket`](@ref), [`ischardev`](@ref), and [`isblockdev`](@ref) can be made on the
@@ -1094,13 +1103,23 @@ returned objects without further stat calls. However, for some filesystems, the 
 cannot be determined without a stat call. In these cases the `rawtype` field of the [`DirEntry`](@ref))
 object will be 0 (`UV_DIRENT_UNKNOWN`) and [`isfile`](@ref) etc. will fall back to a `stat` call.
 
+# Examples
 ```julia
-for obj in _readdirx()
-    isfile(obj) && println("\$(obj.name) is a file with path \$(path(obj))")
+for entry in readdir(DirEntry, ".")
+    if isfile(entry)
+        println("\$(entry.name) is a file with path \$(path(entry))")
+        continue
+    end
+    isdir(entry) || continue
+    for entry2 in readdir(entry)
+        ...
+    end
 end
 ```
 """
-_readdirx(dir::AbstractString=pwd(); sort::Bool=true) = _readdir(dir; return_objects=true, sort)::Vector{DirEntry}
+readdir(::Type{DirEntry}, dir::AbstractString=pwd(); sort::Bool=true) = _readdir(dir; return_objects=true, sort)::Vector{DirEntry}
+readdir(::Type{DirEntry}, entry::DirEntry; sort::Bool=true) = readdir(entry; sort)::Vector{DirEntry}
+readdir(entry::DirEntry; sort::Bool=true) = readdir(DirEntry, path(entry); sort)::Vector{DirEntry}
 
 function _readdir(dir::AbstractString; return_objects::Bool=false, join::Bool=false, sort::Bool=true)
     # Allocate space for uv_fs_t struct
@@ -1202,7 +1221,7 @@ function _walkdir(chnl, path, topdown, follow_symlinks, onerror)
             end
             return
         end
-    entries = tryf(_readdirx, path)
+    entries = tryf(p -> readdir(DirEntry, p), path)
     entries === nothing && return
     dirs = Vector{String}()
     files = Vector{String}()